Rename symbol files
