+++ src/declarations/es-declared.symbol.ts
