+++ src/declarations/mod.ts
@@ -1,3 +1,3 @@
 export * from './es-const.js';
 export * from './es-declarations.js';
-export * from './es-declared-symbol.js';
+export * from './es-declared.symbol.js';
