+++ src/symbols/es-import.ts
@@ -1,5 +1,5 @@
 import { EsExternalModule } from './es-external-module.js';
-import { EsImportInit, EsImportedSymbol } from './es-imported-symbol.js';
+import { EsImportInit, EsImportedSymbol } from './es-imported.symbol.js';
 
 /**
  * Creates a symbol imported from the named {@link EsExternalModule external module}.
