+++ src/symbols/es-imported.symbol.ts
