+++ src/symbols/mod.ts
@@ -1,6 +1,6 @@
 export * from './es-external-module.js';
 export * from './es-import.js';
-export * from './es-imported-symbol.js';
+export * from './es-imported.symbol.js';
 export * from './es-imports.js';
 export * from './es-module.js';
 export * from './es-namespace.js';
